+++ src/matchAccepted/acceptMatch.ts
@@ -1,23 +1,27 @@
-import { fetchSupportRequestAndVolunteer } from "../utils";
+import type {
+  MatchConfirmations,
+  SupportRequests,
+  Volunteers,
+} from "@prisma/client";
 import {
   authenticateMatch,
   confirmMatchConfirmation,
   createMatch,
-  fetchMatchConfirmation,
-  getMatchConfirmationId,
   updateTicketWithConfirmation,
 } from "./matchAcceptedLogic";
 
-export default async function acceptMatch(buttonPayload: string) {
-  const matchConfirmationId = getMatchConfirmationId(buttonPayload);
-
-  const matchConfirmation = await fetchMatchConfirmation(matchConfirmationId);
-
-  const { supportRequest, volunteer } = await fetchSupportRequestAndVolunteer(
-    matchConfirmation.supportRequestId,
-    matchConfirmation.volunteerId
-  );
-
+export default async function acceptMatch(
+  matchConfirmation: Pick<
+    MatchConfirmations,
+    | "matchConfirmationId"
+    | "supportRequestId"
+    | "volunteerId"
+    | "matchType"
+    | "matchStage"
+  >,
+  supportRequest: Pick<SupportRequests, "zendeskTicketId">,
+  volunteer: Pick<Volunteers, "firstName" | "zendeskUserId">
+) {
   await updateTicketWithConfirmation(supportRequest.zendeskTicketId, volunteer);
 
   const authToken = await authenticateMatch();
